pyproject.toml
[tool.poetry]
name = "django-upload-with-base64"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "django_upload_with_base64"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
